# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/products:
    post:
      operationId: ProductsController_createProduct
      parameters: []
      requestBody:
        required: true
        description: Create a product with images, variants, and collections.
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Nike Air Max 2024
                description:
                  type: string
                  example: Stylish running shoes
                categoryId:
                  type: string
                  format: uuid
                collectionIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                discount:
                  type: number
                  example: 10
                variants:
                  type: string
                  example: >-
                    [{"sku":"NIKE-001","size":"42","color":"Black","stock":20,"price":75000}]
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: ''
      security: &ref_7
        - bearer: []
      tags: &ref_8
        - Products
    get:
      operationId: ProductsController_getProducts
      parameters:
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            example: 10
            type: number
        - name: search
          required: false
          in: query
          schema:
            example: shoes
            type: string
        - name: categoryId
          required: false
          in: query
          schema:
            type: string
        - name: collectionId
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_7
      tags: *ref_8
    delete:
      operationId: ProductsController_deleteProducts
      parameters:
        - name: ids
          required: true
          in: query
          schema:
            type: string
        - name: permanently
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_7
      tags: *ref_8
  /api/products/{id}:
    get:
      operationId: ProductsController_getProductById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_7
      tags: *ref_8
    patch:
      operationId: ProductsController_updateProduct
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_7
      tags: *ref_8
  /api/products/{id}/variants:
    get:
      operationId: ProductsController_getVariants
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_7
      tags: *ref_8
  /api/products/{productId}/variants:
    post:
      operationId: ProductsController_addVariantToProduct
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      security: *ref_7
      tags: *ref_8
  /api/products/{productId}/variants/{variantId}:
    patch:
      operationId: ProductsController_updateProductVariant
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
        - name: variantId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_7
      tags: *ref_8
    delete:
      operationId: ProductsController_deleteVariant
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
        - name: variantId
          required: true
          in: path
          schema:
            type: string
        - name: permanently
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_7
      tags: *ref_8
...
  /api/cart/add:
    post:
      operationId: CartController_addToCart
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToCartDto'
      responses:
        '201':
          description: ''
      tags: &ref_13
        - Cart
  /api/cart/remove:
    delete:
      operationId: CartController_removeItems
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveCartItemDto'
      responses:
        '200':
          description: ''
      tags: *ref_13
...
  /api/orders:
    post:
      operationId: OrdersController_createOrder
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: ''
      tags: &ref_14
        - Orders
  /api/orders/{id}:
    get:
      operationId: OrdersController_getOrder
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_14
    delete:
      operationId: OrdersController_deleteOrder
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_14
  /api/orders/user/{userId}:
    get:
      operationId: OrdersController_getOrdersByUser
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_14
  /api/orders/{id}/status:
    put:
      operationId: OrdersController_updateOrderStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusDto'
      responses:
        '200':
          description: ''
      tags: *ref_14
info:
  title: E-Commerce Fashion API
  description: API documentation for the Fashion E-commerce backend
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    AddToCartDto:
      type: object
      properties: {}
    RemoveCartItemDto:
      type: object
      properties: {}
    CreateOrderDto:
      type: object
      properties: {}
    UpdateOrderStatusDto:
      type: object
      properties: {}